Adding an article to week 02 of articles of the week

+++ myadmin/views.py
@@ -13,7 +13,7 @@ from django.views import generic
 from django.views.decorators.http import require_http_methods
 from django.views.generic import TemplateView
 
-from comment.models import Comment
+from comment.models import Comment, Danmaku
 from helpers import get_page_list, AdminUserRequiredMixin, ajax_required, SuperUserRequiredMixin, send_html_email
 from users.models import User, Feedback
 from video.models import Video, Classification
@@ -59,13 +59,17 @@ class IndexView(AdminUserRequiredMixin, generic.View):
         user_today_count = User.objects.exclude(date_joined__lt=datetime.date.today()).count()
         comment_count = Comment.objects.get_count()
         comment_today_count = Comment.objects.get_today_count()
+        danmaku_count = Danmaku.objects.get_count()
+        danmaku_today_count = Danmaku.objects.get_today_count()
         data = {"video_count": video_count,
                 "video_has_published_count": video_has_published_count,
                 "video_not_published_count": video_not_published_count,
                 "user_count": user_count,
                 "user_today_count": user_today_count,
                 "comment_count": comment_count,
-                "comment_today_count": comment_today_count}
+                "comment_today_count": comment_today_count,
+                "danmaku_count": danmaku_count,
+                "danmaku_today_count": danmaku_today_count}
         return render(self.request, 'myadmin/index.html', data)
 
 
@@ -242,6 +246,36 @@ def comment_delete(request):
         return JsonResponse({"code": 1, "msg": "无删除权限"})
     comment_id = request.POST['comment_id']
     instance = Comment.objects.get(id=comment_id)
+    instance.delete()
+    return JsonResponse({"code": 0, "msg": "success"})
+
+class DanmakuListView(AdminUserRequiredMixin, generic.ListView):
+    model = Danmaku
+    template_name = 'myadmin/danmaku_list.html'
+    context_object_name = 'danmaku_list'
+    paginate_by = 10
+    q = ''
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super(DanmakuListView, self).get_context_data(**kwargs)
+        paginator = context.get('paginator')
+        page = context.get('page_obj')
+        page_list = get_page_list(paginator, page)
+        context['page_list'] = page_list
+        context['q'] = self.q
+        return context
+
+    def get_queryset(self):
+        self.q = self.request.GET.get("q", "")
+        return Danmaku.objects.filter(content__contains=self.q).order_by('-timestamp')
+
+@ajax_required
+@require_http_methods(["POST"])
+def danmaku_delete(request):
+    if not request.user.is_superuser:
+        return JsonResponse({"code": 1, "msg": "无删除权限"})
+    danmaku_id = request.POST['danmaku_id']
+    instance = Danmaku.objects.get(id=danmaku_id)
     instance.delete()
     return JsonResponse({"code": 0, "msg": "success"})
